LayoutTests/imported/w3c/web-platform-tests/focus/hasfocus-same-site-expected.txt


FAIL Check result assert_equals: Check result expected "PASS" but got "FAIL: focus not lost when other window opened"
